requirements.txt
# FastAPI y servidor
fastapi==0.109.1
uvicorn[standard]==0.24.0
python-multipart==0.0.20

# Pydantic para validación
pydantic==2.12.1
pydantic-settings==2.11.0

# LangChain - versiones compatibles (deja que pip resuelva)
langchain>=0.3.18
langchain-community>=0.3.18
langchain-core>=0.3.79
langchain-text-splitters>=0.3.6

# LLM: Groq (gratis)
langchain-groq==0.3.8

# Embeddings: HuggingFace (local, 100% gratis)
langchain-huggingface==0.3.1
sentence-transformers>=2.6.0

# Vector DB: pgvector (requiere Python 3.11)
psycopg2-binary==2.9.11
pgvector==0.2.5
asyncpg==0.29.0

# Database ORM y migraciones
sqlalchemy==2.0.25
alembic==1.13.1
greenlet==3.0.3

# GCP (solo para Cloud Storage y autenticación)
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.16.4

# Utilidades
PyYAML==6.0.1
python-dotenv==1.0.0
python-dateutil==2.8.2

# HTTP client
httpx==0.25.2

# Rate Limiting (protección anti-DoS)
slowapi==0.1.9

# Testing dependencies
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
